refactor: split utils.ts
